HomeWork8/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
trl>=0.7.0

# Evaluation metrics
evaluate>=0.4.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# PDF processing
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
Pillow>=9.5.0

# Web scraping and data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0

# Optional: For quantization and optimization
bitsandbytes>=0.41.0
peft>=0.4.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and monitoring
tensorboard>=2.13.0
mlflow>=2.5.0

# Additional utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0

